refactor jUSDC oracle with chainlink
